requirements.txt
langchain-iris
testcontainers-iris
llama-iris
llama-index-legacy
google-genai
dotenv
google-cloud-aiplatform
fastapi
uvicorn
python-dotenv
sentence-transformers
pydantic
